Refactor: Plural enums, dictionaries
Fixed dictionaries to be consistently named '_foos_by_bar', with the
first noun plural and the second noun singular. There are a few
exceptions where the dictionary's key is actually a hybrid key with
several different parts, in which case it actually should be plural.

Fixed enums to be consistently singular with names like 'State' and
'TileType' instead of 'States' and 'TileTypes'