Refactor: min_volume check, post_unemote
Made min_volume check consistent; instead of checking if two flats are
equal or if their difference is within a threshold, we now use Godot's
built-in 'is_equal_approx'

EmotePlayer no longer uses 'yield' to wait for a tween to finish. It now
uses tween's tween_callback functionality. This avoids an error if the
scene is disposed while yielding.