Moving no longer interrupts mood. Disabled north-facing emotes.
Instead of interrupting the movement animations, moving just resets the
creature's "emote arm frame". This is the same mechanism that happens
when you launch an emote while the creature is running.

Creatures will no longer emote when facing north. This was resulting in
some buggy behavior like eyes on the back of their head.

Closes #1905.

Added movement options to MoodDemo.