Added 'spear' asparagus critter.
Calling it 'asparagus' is a lot of typing, and asparagus has no plural
form in English. I've settled on 'spear/spears' which is less typing and
more grammatically sensible, so we don't end up awkwardly trying to
explain things like 'asparagus is a list of asparagus'

I experimented with having spears cause line clears immediately when
they pop out, but it caused some problems. LineClearer does not support
the idea of clearing more lines when lines are already being cleared, so
there were bugs where not all lines got cleared correctly.

Spears are the first obstacle which highlight '!' when they're JUST
about to pop out. I think this is important because their behavior is so
disruptive, and it's no fun to lose to an unwelcome surprise. (Well,
maybe it's a little funny the first time.)