Replaced level timer 'initial_interval' with 'start/end'
Level timers used to have an optional 'initial_interval' property to define when
the timer started, but there was no way to stop a level timer. I've replaced
this 'initial_interval' property with a 'start' property, which is more
intuitive and matches the naming conventions of Godot's timers. I've also added
a new 'end' property.