FrameOverworldObstacle improvements: shuffle_frames, initialize
Some upcoming obstacles have animations and static images in the same
sprite sheet, so we might only want to allow an arbitray set of frames
when shuffling.

FrameOverworldObstacle now initializes its onready variables if it
detects its sprite is null in the editor. This happens sometimes when
editing a tool script.