Fixed shark crashes related to PieceWritten trigger, piece orientation
The PieceWritten trigger fires when a shark eats a piece. This trigger
can cause level effects like inserting lines, and moving sharks, which
then was causing crashes (for sharks that could not be found) or
graphical oddities (for sharks that moved)

The PieceWritten trigger now only fires after all sharks eat.

Eating a piece with a non-zero orientation was causing crashes in some
cases. I've moved the check to 'UpdatePieceManagerPiece' so that a null
piece will never have a non-zero orientation.