Fixed rainbow lights for cake clears.
Clearing a cake box results in rainbow lights appearing, but the rainbow
included an olive green vegetable color when all the tile indexes were
shifted in 86fed88a.

I've fixed the index problem and refactored the code so that further
changes to the FOOD_LIGHT_COLORS array will not break this code.