New customers: Kylie, Jaco, unknown
Jaco will appear during the credits sequence.

'unknown' is used for cutscenes which need things like anonymous voices.
A nametag is not shown for this creature, and the creature themselves
should never really be spawned anywhere. We just need them for their
chat theme.

Kylie is just a random customer.