Fixed 'singleton environment' dependencies
A lot of environment code for things like shadows and cameras relied on
the singleton CreatureManager to get the creatures in the scene. For the
upcoming credits scenes, this no longer works because we might have a
cutscene playing with a camera, and an unrelated creature pinup
scrolling by. The existing logic would try to zoom out the environment
camera to include creatures who weren't even in the scene.

I've changed code for things like 'get all creatures' or 'get the
creature with this ID' to require an environment parameter. This ensures
logic for things like cameras and shadows only affects the creatures
within that environment.