Bug: Polygon error with unusual fatness values
When using pinup-demo.gd and changing the creature size from 10 to 1 abruptly, an error is logged.

    E 0:00:05.836   canvas_item_add_polygon: Invalid polygon data, triangulation failed.
      <C++ Error>   Condition "indices.empty()" is true.
      <C++ Source>  servers/visual/visual_server_canvas.cpp:768 @ canvas_item_add_polygon()

This specific error was because CreatureCurve lerps its polygon geometry
based on the VisualFatness value. When it goes outside a specific range,
it continues extrapolating in that direction but this eventually results
in self-intersecting polygons.

We could fix CreatureCurve to avoid this use case, but even so -- values
higher than 10.0 will result in oddities because the heads and arms are
not lerped in the same way curves are, so they will stay constrained
while the body grows bigger and bigger. We should just avoid setting
these kinds of values.

Closes #2361.