Added credits pinup transformations.
Orb now fires some of its pieces towards pinups, which are then
transformed into silly anime caricatures of the characters.

Tweaked AnimationPlayer sync; Syncing the AnimationPlayer to 4.100 FPS works
better with the music, which is 123 beats per minute. It's still not perfectly
in sync with the music but it's as close as I think I can get.