Crowd cutscene refactoring: played, _on_Director, unused tween
Renamed 'creatures_arranged' signal to more generic 'played'. A future
expansion of this method will add new methods and signals named 'stop'
and 'stopped', so this is more consistent and simple.

Renamed '_on_CrowdWalkDirector_creatures_arranged' signal to
'_on_Director_creatures_arranged'. This signal method is used for the
CrowdSurfDirector as well, so the name should be generic.

Removed unused CrowdSurfDirector._tween field.