Chocolava/credits applause SFX are continuous
Replaced 6-second cheer sound effect with a continuous sound effect.

The cutscene now specifies when it wants the sound effect to stop with a
'stop_sfx' call. I also removed the second 'lava_cheer_2' sound effect.
This was a slightly different sound effect with more lead-in applause,
but it felt unnecessary.

The credits now play applause continuously, and the volume decreases as
the crowdies stop cheering. There is a second separate applause sound
effect for the lead-in applause when tossing the characters in the air.