Improved pinups scrolling offscreen during credits.
One especially large pinup 'chelle' was blinking out of view
unpleasantly. Pinups now slowly fade out after they're above a certain
height. This also improves the appearance when the window is very tall,
as the pinups fade in from the bottom and fade out at the top, so it's
symmetrical. Before, the pinups kept scrolling up all the way offscreen
and it looked a bit strange.