Export, packaging configuration for steam.
package.sh script packages '-steam' folders, includes .dll and .so files
needed for GodotSteam.

Removed HTML5, Android exports. These have nothing to do with steam, and
should be exported from the main branch.

Removed iOS export. TurboFat is not playable on iOS, and configuring it
for Steam requires some additional steps that did not seem trivial.

Added custom GodotSteam templates to Linux, Windows. The GodotSteam
Exporting and Shipping tutorial states:

> 'Since the pre-compiles are never debug-release versions, you must make sure
> that only the release field contains a template and that the debug field is
> empty... [snip] ...If you have the debug field filled in, especially while
> using the pre-compiled templates, the export process will fail'.

This appears to be untrue on two accounts; GodotSteam now provides debug
versions, and leaving the debug filled in while using the pre-compiled templates
does not cause the export process to fail. The debug version exports fine. I'm
leaving the debug field filled in but we should watch for symptoms if something
is wrong with the exported version.