Removed unused 'ui_has_focus' property.
This was used in FreeRoam mode when the player could move with the arrow
keys and talk to other creatures. When the dialog boxes popped up, we
didn't want them moving around when trying to select chat choices. This
functionality is no longer in the game, not even in demos, so we don't
need this property anymore.