Move CandyButton signals from .tscn to script.
Moved CandyButton signals so they no longer appear in the editor. This is a
generic button used in many places, we want to be able to quickly see the unique
signals connected to each button instance, not the generic signals connected to
all button instances.

Assigned more suitable default textures/color to CandyButtonH4.