Vega Churn level difficulty tweaks. PlayerData '55aa'.
Adjusted difficulty and duration for levels formerly in Vega Churn
Twelve. For most levels, this was just tweaking a few numbers but a few
levels involved more substantial changes:

'Blindy Asparagus': Asparagus now pops out earlier and more
aggressively, to compensate for the shorter level duration.

'Carrot Cake 2': Carrots now appear earlier to compensate for the shorter level
duration.

'Checkerboard Vault': Less digging required to access the vault, and a
greater reward for accessing the vault. This is partially to compensate
for moving it to an easier world, and partially to compensate for how
difficult it was to score well.

'Secret Brownies': No speedups and fewer starting pieces, to compensate
for the shorter duration.

'Snack Box Blitz': Now uses Merrymellow Marsh's rotating next queue
gimmick.

'Strawberry Slowcake': Now uses Poki Desert's shark gimmick. Renamed to
Strawberry Sharkcake.

Changed PlayerData version to '55aa'. During the upgrade, we remove
scores for the levels which have been adjusted. This is to avoid the
scenario where a player's level history has an unreachable score for a
level which was shortened.