Redesigned creature editor for gamepads, non-default UI
Redesigned the creature editor. The new design avoids use of the default
grey Godot UI, and default widgets such as sliders and text boxes. It is
also very gamepad friendly. It does not require scrolling, and
categories can be navigated using the bumpers.

The old creature editor is still accessible via 'CreatureEditorOld' in
the demo.

There is still some work remaining for the new creature editor:
 - name entry still requires keyboard
 - the color picker interface does not support gamepads
 - the old 'mutate' functionality was fun, and would be nice to keep
   around in some form
 - chat themes can still not be edited, and the constant randomization
   of them makes the 'body color picker' look a little strange as the
   nametag is constantly flickering