Support tab navigation with gamepad bumpers.
Tab navigation is now supported with Z, X, or the gamepad bumpers. We
also continue to support arrowing up into the TabContainer, but the
gamepad bumpers are a bit faster.

The real reason for implementing this is that the Creature Editor
redesign will use a similar tabbed approach, and I'd like to streamline
it with bumpers.

Closes #2526.