Refactor: Simplified default dialog path behavior.
Eliminated 'assign_default_dialog_path' method. The purpose of this
method was to detect whether a dialog had been launched or not, and if
not, to assign it a default path. It is much simpler to instead assign the
default path in the _ready() function. It no longer requires a
specialized function or weird UUID path.

Changed level editor and old creature editor from defaulting to 'C:\' or
'<snip>/AppData/Roaming/Godot/app_userdata/Turbo Fat' to 'Documents'

Removed redundant level editor 'OpenResourceDialog.popup_centered' call.
This was being called twice by mistake.