Bug: Gut tests were deleting user data.
Gut tests had logic to override the player's data_filename, but this was
then being reverted when the SystemSave reacted to the 'save_slot_changed'
signal, setting the filename back to the user's save slots. The result
was that some tests would delete the user's data instead of deleting the
test's data.

I've added two protections for the Gut tests:
1. SystemSave no longer assigns the filenames during tests.
2. Gut tests override the filenames by default in a pre-run hook.