CreatureEditor's color picker supports gamepads
Created 'CandyColorPicker', a color picker which supports gamepads. It
includes a context-sensitive set of color presets as well as HSV
sliders for fully custom colors. It can be navigated well using a
gamepad or keyboard.

The presets were collected from both the presets from DnaUtils as well
as the creature colors for creatures in the game (e.g Rhonk, Skins).
Utilities for repeating this process are in CreatureColorButtonDemo.