Removed level editor from main menu.
This editor was removed because it used undecorated UI, and it was also
unintuitive, incomplete and incompatible with gamepads. It will need a massive
overhaul before it can meet the standards to be included in the game.

The level editor is still available in the demo folder,
'src/demo/editor/puzzle/LevelEditor.tscn'.