Redesigned TrainingMenu with LevelButtonScroller
The training menu previously used an undecorated button which was very
wide, and would have been difficult to reskin as a candy button given
its width.

The training menu now uses a 'LevelButtonScroller' which uses the same
button style as the career mode and level picker. We'd like to
eventually amp up these buttons to include some icons or fancy graphics,
so that will help this screen look a bit prettier too.

The training menu's high score table now uses the 'squeak-theme-gy'
variant, instead of an undecorated 'h4' theme.