Fixed 'unpickable boss level' bug.
These boss levels would usually be pickable because the player would play them
to advance. But if the player skips one of these with a cheat code they
could never be played, because their unlock criteria was 'play the
level' and the criteria for playing them was 'you haven't cleared this
area'.

Boss levels now have an additional unlock criteria of 'you've cleared
the region'. If the level is skipped, they'll be unlocked anyways.

Closes #2364.