Retired 'a little garbage'
This level wasn't very fun. Digging in Turbo Fat is extremely difficult,
and while other levels like 'A Little Garbage 2' and 'Sharp Curds'
mitigate this by having pleasant patterns to dig through and
high-scoring pickups, this level is low-scoring and very difficult to
dig through.

I experimented with a few ways of salvaging the level (higher line
scores, holes in just 3-6 columns of the level) but I couldn't make it
an interesting level.