Refactor: Extracted GradeLabels.tscn
Extracted GradeLabels node. The script was already reused, but the
node encapsulates the 'HookableLevelGradeLabel' packed scene, mouse filter, anchors
and other node properties.

Removed LevelGradeLabels 'button_added' signal receiver. This was a bit
of a 'gotcha' when extracting this scene, and I don't like the design.
This signal method is only used by PagedLevelPanel, and it makes more
sense to be on one of PagedLevelPanel's scripts instead.