Bug: Black (#000000) creature colors caused texture issues
CreatureLoader treated 'black' as 'null' when assigning shader colors.
Black is a pickable eye color in the creature editor, and the player can
also pick it with the sliders.

'Transparent' is now treated as null instead. The player can't pick this
with the creature editor, so it's more suitable as a 'null' value.