RankCalculator.rank() -> average_rank_for_grade()
The name 'rank' was ambiguous, and a reasonable misinterpretation of the
function is that it would convert a grade like 'A' to the minimum
requirement such as '44.0'.

I've renamed the method to 'average_rank_per_grade' and added a new
method 'required_rank_for_grade' to provide the utility I needed.