Bug: "winish" cheat caused bizarre output for ResultsHud
Toggling the "winish" or "finish" cheats adjusts the player's score, but
didn't adjust their score components. The result is that the ResultsHud
would count up a total of $0 and then award an SS+ grade, which makes no
sense.

The "winish" and "finish" cheats now make up arbitrary values for the
combined scores.

Also made a minor adjustment to ResultsHudBlueprint to explicitly
calculate the 'extra' value as 'everything but boxes and combos'. This
helps handle edge cases where they don't add up otherwise. It also helps
future-proof the code in case the score calculation changes in the
future.