'Level failed...' when not scoring enough for a boss level.
ResultsHud now shows 'Level Failed...' when failing to meet the
success criteria for a boss level. Before, it only showed a failure
message if you quit or lost all your lives.

Added 'CurrentLevel.boss_level' flag. This is necessary because the
CareerData.is_boss_level() method is based purely on the player's
distance, but this distance is unreliable during a level, as it is
adjusted when the player plays or finishes a level