Added instance_valid check to avoid NPE in fatness changer
When importing a creature, this code occasionally crashed complaining
that a method was being called on a null instance. The instance wasn't
actually null, but was not a valid instance. This check seems to prevent
the error.

The error didn't happen very consistently, but seemed to happen if you
played the game for awhile and then launched the creature editor and
imported a creature.