'One Step Ahead' bugs: Spin move noise sfx, unrotateable pieces
Fixed bug in 'One Step Ahead' where spin move sfx would play if the
piece was locked in place, where it required a spin move.

Fixed bug in 'One Step Ahead' where the next piece would not rotate if
the piece was locked in place, where it could not rotate.