Added goal hype music: 'Extra Sprinkles'
'Extra sprinkles' plays for boss levels. It plays once the player
reaches the boss level, and under certain other edge cases like if the
player quits and re-enters career mode when they're at a boss level.

Removed 'Extra Sprinkles' as Chocolava Canyon's signature song. Sugar
Crash is now the signature song for the first two regions, and Pressure
Cooker is the signature song for Chocolava Canyon. This is good for a
few reasons imho: Sugar Crash shows up in the credits, so it's good to
cement it in the player's head a little more. Also Extra Sprinkles is a
really good song so it's good for the player to hear it early, as
opposed to saving it for levels which only a few players will reach.