Remove savescum prevention code.
It's not obvious when the player is "allowed" to quit, and triggering
this code results in an unintuitive occurrence where you pick career
mode, it shows you the day is over, and dumps you back to the main menu.

If the player wants to save scum, let them save scum. It is a
single-player game.