Bug: Tutorial replays broke if player held a key
If the player held an input before a level transition, the tutorial
replays would break. It would ignore all player input, but this ignored
the "releasing the input" input so the entire replay would play as
though the key were held.

All keys are now released at the start of a level transition.