Improve quit behavior during credits
Before, watching the final cutscene would end the career day, but if you
exited the previous cutscenes, the player would continue into Starberry
Mountain.

This isn't a big deal, but it's a little inconsistent and odd. The
gameplay behavior should stay the same regardless of whether the player
watches the credits or the preceding cutscenes.

I also closed a small issue where watching the cutscenes in the
CutsceneDemo (or perhaps some future gallery mode) would end the current
career day. The career day is now only ended if the player is watching
the cutscenes from career mode.