Added level intro goal message, 'Earn ¥25 with 15 lines...'
Added a goal message which shows up before non-tutorial levels. This
level serves two purposes. For normal levels, it gives the player an
idea of what score they should be aiming for either for a baseline rank
or to reach the next rank milestone. For boss levels, it tells the
player the necessary criteria to progress.