Improved hint messages: cheats, soft drop lock cancel
Added hint messages for turning on cheats. Some players are going to be
discouraged at the game's difficulty, so it would be nice to
occasionally remind them about the cheats they can turn on.

Changed 'soft drop lock cancel' hint to tell players where to enable
this functionality. It's off by default, they'll have to turn it on.