Set up a semaphore system for get_tree().paused
Multiple sections of code call get_tree().paused to pause and unpause
the game, but they could conflict. For example, a player could open the
settings menu and the steam overlay, and then close the steam overlay.
Right now the game unpauses, but it should stay paused.

This new semaphore system ensures the game remains paused if any one
script wants it to be paused.