Bug: buffered player inputs impacted replays
If the player mashed the rotate keys when a replay was starting, their
buffered input would be applied to the replay. For example, during the
'squish and recover' tutorial, they could mash the rotate buttons during
the two replays of the tutorial, causing Fat Sensei to clear a line or
to block the well entirely.

Buffered inputs no longer impact replays.