Minor display issues with vertical screens.
Cutscene/chat UI now extends all the way to bottom of the screen.
Before, chat choices and chat messages were appearing in the middle of
the screen.

Progress board is now centered, instead of appearing at the top of the
career map window.

Training menu speed slider is now centered.