Clip results hud receipt vertically.
The results hud receipt is now clipped. Its boundaries extend 84 pixels
below the bottom of the screen. This distance means it will remain
unclipped for users playing the game at a 4:3 aspect ratio, but also
means it won't extend so far as to look too weird on a vertical monitor.