Bug: Restarting sometimes kept focus on current 'morphing' creature
When restarting a puzzle with a priority customer, the camera would
often stay focused on that customer when the player broke their combo.
This meant instead of swinging to a new customer and having the old
customer change in the background, the camera would stay focused on the
current customer and the customer would change in the foreground.

I've changed the 'puzzle preparation' code to always purge the current
customer from the queue to ensure it will always focus on a new
customer.

I also changed RestaurantView.reset() to kill all the
summon_customer_timers. If the timers stayed running, there were some
edge cases where the camera would end up on a random customer instead of
the intended priority customer.