Calculate rank data at runtime, instead of storing it
Rank data is calculated at runtime instead of being stored. Storing rank
data had one major design problem, which is that if we want to make a
level easier, the store rank data will still give the player a worse
grade

We also purge all non-existent levels.

Calculating all rank data takes about 400 ms on my computer, and it's
only done at startup.

If we decide the 400 ms delay on startup is undesirable, we could adopt
a 'best of both worlds' approach where rank data is stored, but
conditionally refreshed (perhaps the player's save data can have a
'level_version', which causes us to recalculate all ranks when it's out
of date, or something like that.)