Fixed gut tests for rank change
Now that ranks are populated on startup, it takes about 300 ms to load a
save. This is fine for the player starting the game, but it's too slow
for unit tests since they need to load 20-30 saves to test backwards
compatibility. I've added a 'pre-run hook' which disables this
functionality. It invalidates 2 tests.