Protect against unusual 'getAchievement' responses
We're seeing crashes seemingly centered around calls to
`getAchievement`. It's unclear what's causing this, but one possibility
is that getAchievement is not returning a Dictionary with a boolean
'achieved' key -- if any of those weren't true, this method would crash.

I've added some logging and checks to protect against unusual responses.