Notify the player when changing languages.
When changing languages, the UI will look a little strange, because some
UI details are initialized once, and not reinitialized in response to
the language changing.

Other games such as Slay The Spire tell the player 'Restart the game for
display changes to take effect', but the game seems to work OK as long as the
player changes scenes (starts a level, enters career mode, etc.) If we
later find bugs where the player really needs to restart the game, we
can change the message.