Removed redundant spanish translation of data objects
Our data objects such as 'CareerRegion' and 'LevelSettings' translated
their name and description on initialization. This was a flawed idea for
many reasons:

1. These models are only initialized on startup, which means if the
   objects were translated into Spanish on startup, the player would not
   be able to change their language as our translation framework has no way
   of translating from Spanish to other languages

2. These models are initialized before PlayerSave is loaded, so they
   were always translating into English anyways, even if the player
   selected a different language

Overall it seems better for the models to remain in English, and for
language translation to occur on the UI side. I've verified that career
regions and levels still show up in Spanish, even removing this
translation layer.