Added timeout for mouse hiding logic.
One user reported that when moving the mouse, the mouse flickered
constantly. This could happen if they had a joystick leaning on its
side, or some other anomaly which constantly triggered
InputEventKey/InputEventJoypadButton events. I've added a timeout so
that the mouse will stay visible for 0.5s after the most recent mouse
event, even if there is keyboard/joypad input.

Closes #2828.