Streamline achievement checks.
Achievement conditions do not need to be checked if the achievement has
already been achieved. We used to avoid calls to
Steam.is_achievement_achieved because API calls were expensive,
resulting in internet calls and sometimes crashing the game. However
these calls are now cheap and go to our facade, so it's better to avoid
checking the achievement conditions for completed achievements. These
achievement checks in some cases are a little expensive, such as for the
'Create Creature Achievement'.