Added null checks for CurrentLevel.puzzle.
If certain signals were fired while no puzzle was activate, a null
pointer exception could cause a crash. I never saw this myself, as it
would require something nonsensical like "a shark is smushed on the main
menu screen," but it could potentially explain some of the crashes we're
seeing on steam versions of the game.