Added 'troubleshoot.bat' file, verbose logging
Turbo Fat now includes a 'turbofat-troubleshoot.bat' file which enables
Godot's '--verbose' flag as well as a Turbo-Fat-specific '--tf-verbose'
flag. These enable enhanced logging for scene transitions, singletons
and resource loading. The intent is to give players a tool to share
detailed crash information with developers, to try and root out the
cause of the game crashing after puzzles.