Save indicator appears sooner.
We used to show the save indicator at the start of the save operation.
However because our save logic is non-threaded, this meant the save
indicator would not be shown until the save operation was complete.

I've added a new signal so the save indicator is shown before the scene
transition if a save is scheduled.