'Cake tutorial' no longer requires squishes/spins.
I personally thought it was cute how the cake tutorial shuffled the
pieces every time, and how it was always possible to complete regardless
of the piece order. However based on player feedback, this tutorial was
too difficult and the shuffling confused some people, or they were
frustrated that they had to use techniques such as spin moves or squish
moves to complete it.