Never start Adventure mode with 'You're done!'
Currently, if the player gives up on Adventure Mode or runs out of lives and
quits to the main menu, then the next time they start Adventure Mode it shows
them "Your day is over, here's how you did" and dumps them back to the main
menu.

I've changed this so that it seamlessly advances them to the next day.

Closes #2848