Chalkboard font tweaks: use_filter, line_spacing
Enabled 'use_filter' for chalkboard fonts. We enable 'use_filter' for
our fonts elsewhere so that they scale and rotate more smoothly. I
wanted to rotate the chalkboard font in a screen I was designing, but it
lookde ugly. It still looked ugly with this setting, but I'd still
prefer to keep our font settings consistent.

Decrased 'line_spacing' from 3 to -6. The chalkboard font is very tall,
and having too much space makes it look strange. It is currently only
used on the Adventure mode win screen for single-line messagess, so this
does not affect the appearance anywhere, but it is used in the upcoming
Adventure mode intro.