Easier spin tutorial, more demonstrations.
The spin tutorial now gives tutorials if the player fails 4 times.

The piece flip part of the tutorial also shows the exact steps the
player should follow, instead of showing them how to flip a piece off to
the side in an unnecessary place. This confused a playtester a few years
ago, and while it's an interesting puzzle, it's out of place in a
tutorial