Let player boss level, even if time is up
Before, if you reached the boss level on your sixth level, there would
be a big fanfare but you wouldn't get to play it.

Now, you are always allowed to play the boss level even if you are out
of time.

I've also fixed it where the fanfare does not play if you are not able
to play the boss level -- for example, if you lose all your lives, but
advance two spaces and land on a boss level. Before, this would result
in a big fanfare with boss music and then the day would end. Now, there
is no fanfare or special music.

Closes #2873.