Bug: Don't show ending cutscenes when retrying levels
In Adventure mode, retrying a level counts as a failure and shouldn't
display cutscenes. However, our logic for 'do we play a cutscene' was
keying off of the player's best performance, not their first
performance.

While it's not necessarily a bad thing to be lenient and say, 'oh it
took you a few tries, but here is a cutscene', it was inconsistent with
how Adventure mode works. It meant a player could lose a boss level, retry, beat
it, see a cutscene, and then get sent backwards on the map screen and
told 'try again tomorrow!'

Retrying a level now prevents cutscenes from playing. The cutscene logic
cares about your first performance, not your best performance