Bug: Some level results were being saved twice
In Career Mode, restarting a level would record the loss to your level
history twice. One save was caused by PuzzleState's end game triggers,
and one save was directly embedded in the 'restart' logic.

I've added a conditional so that the 'restart' logic does not record the
loss during the first level in Career mode. It is a little kludgy but I
can't think of a better fix at the moment.