Persist save data after upgrading it.
Before, after upgrading old save data, the old data would remain on disk
and need to be upgraded every time the player saved the game. This isn't
a big deal, since newer versions are written if the player makes any
change or plays any level. But it's possible a player could configure
their settings, and not touch their settings for 6 months, in which case
the old settings version would stay on their disk until they adjusted
the volume or switched their save slot.

Save data is now immediately persisted after it is loaded.