Upgrade config and save files on startup.
Before, we would upgrade config and save files in memory but not persist
them. This resulted in edge cases where, for example, a player could
have a config file which was running the newest version, one save slot
which was 2 weeks old, and a second save slot which was 2 months old.

This especially introduced complications with the newest release which
moves difficulty settings from the config into the save slots. It meant
that if we wanted to preserve the player's difficulty settings, we'd
need to permanently persist them to the config file just in case 6
months from now they decided to load an old save slot which needed the
saved difficulty settings.

We now upgrade all config files and save files on startup, which gets
rid of these edge cases and lets us streamline the saved data.