Fix difficulty upgrade edge cases: slowest, fast, etc
Difficulties such as 'fast' and 'sloweset' can not be picked in the new
difficulty interface. Before, players who selected these difficulties
would be able to continue using them, although the UI would make it look
like they had no difficulty selected.

Now, when upgrading to the newest version, the player's difficulty is
adjusted to one of the four standard difficulties. They can still hack
into their save files and assign a non-standard difficulty, and the old
behavior will persist. This might be construed as a feature and not a
bug, it doesn't break anything and lets them try the super silly
'fastestest' difficulty if they really want.