Fixed 'null tween' error in 5ed690e1
5ed690e1 changed how nodes are cleaned up during scene changes, causing
an error in Creature's fade_in_started listener, because the listener
was called after the creature was removed from the tree. This
specifically occurred in the 'CutsceneDemo.tscn' but it's possible it
would also affect cutscenes in the game.

We now disconnect the listener when the creature is removed from the
tree.

I regression tested other parts of the game, including the intro,
tutorial, puzzles and credits and did not encounter any errors.