Additional 'null player checks' for 45b49cdb1
While the root cause of 45b49cdb1 was caused by a lingering listener
connected to a node outside of the scene tree, the lines causing the
crash were related to OverworldEnvironment.get_creature_by_id not
handling the case where it is not in a scene tree.

I've changed OverworldEnvironment.get_creature_by_id to return null in
this case, and changed many of its callers to handle this case more
gracefully. This doesn't fix any crash that I know of, but it makes this
code more resilient to future crashes I don't know about yet.