Fix career map 'no scene tree' crash from 5ed690e1
5ed690e1 changed how nodes are cleaned up during scene changes, causing
a crash on the Career Map if you click the 'go back to main menu' button
just before the progress board disappears. These errors are caused by
Breadcrumb's Godot #85692 workaround which delays the freeing of nodes in the
previous scene.

We now disconnect the listener when the career map is removed from the
tree.