Separate 'M' and 'TOP' grades, to allow for lenient master grade
RankCriteria now defines a distinct 'TOP' grade. Currently this is the
exact same as the 'M' grade, but we can now decrease the score/time
requirements for 'M' grades to make the grade more achieveable.

Since levels define their own arbitrary values for what constitutes an
'M' rank, this might be counterintuitive. Why define an arbitrary value
of what a 'TOP' grade is, only to have master grades be an arbitrary
fraction of it? However, keeping these separate allows us to more easily
make sweeping changes such as deciding a master grades is 99% or 80% of
a perfect grade. Otherwise, we'd need to make those changes in 200
different level definitions.